Assortative mating in MCMC.
v0.2

Mothers mate with males of the same flower colour with probability p, and with males of
any other colour with prob 1-p. To do this, I added:

* attribute `matches` to faps_data class: a boolean matrix indicate which mothers match which candidates.
* method `update_assortment_probs` to `faps_data` to add a covariate matric controlling this.
* Functionality to update assortment probabilities within run_MCMC.

In addition, I added function `mcmc.check_parameters` to check parameters passed to run_MCMC can
be evaluated by the function, and to add placeholder values for likelihood, priors and posteriors.

On branch assortment-mcmc
Changes to be committed:
	modified:   002.library/python/amajusmating/faps_data.py
	modified:   002.library/python/amajusmating/mcmc.py
	modified:   002.library/python/setup.py
	modified:   002.library/python/tests/test_faps_data.py
	new file:   002.library/python/tests/test_mcmc.py
	modified:   003.scripts/setup_FAPS_GPS.py
	modified:   005.results/003_mcmc_restrict_kurtosis/chainA.py
	modified:   005.results/006_dispersal_plus_simple_assortment/chainA.py