Covariates update with single function.
Added method `update_covariate_probs(` to `faps_data` objects to take all
the covariate functions and update at once, and insert these into the
paternity array objects. This replaces the corresponding chunk of code in
`run_MCMC`, and will be used in mating functions.

Added a proper test for `run_MCMC` to check this works, and moved test
for `mcmc.check_parameters(` to its own file.

 On branch posterior-mating
 Changes to be committed:
	modified:   002.library/python/amajusmating/faps_data.py
	modified:   002.library/python/amajusmating/mcmc.py
	new file:   002.library/python/tests/test_check_parameters.py
	modified:   002.library/python/tests/test_mcmc.py