MCMC paternity by full integration
I have adapted the MCMC procedure to use `integration='full'`.
This ought to be a better implementation of the function to calculate transition
probabilties, but I previously couldn't get MCMC to work. I think that was
because the proportion of missing fathers was allowed to vary, but there is a
high likelihood solution where everybody's father is missing, and the algorithm
coverged on that. Even strong priors were not enough to stop it.
To address this I now use fixed values for missing fathers, but use a different
value for each MCMC chain (0.1, 0.2, 0.3, and 0.4).

I was also previously starting my chains from random draws from the prior for
each parameter. This took ages to converge, but I still didn't want to change
the priors. Instead, initial conditions are now defined with a list of manually
defined values, which are parsed with a loop. I can reuse this for other results
using different priors without having to create a ton of scripts and risk
copy-paste errors.

During convergence I found that shape and scale parameters sometimes wander
below zero or above one. I had to update `mcmc.update_parameters` so that when
this happens, values are reflected back on themselved. For example:

1. if a value of -0.01 is proposed, this becomes 0.01.
2. If a value of 1.01 is proposed for something bounded by 1, this becomes
    2-1.01 = 0.99.

Also updated MCMC log files to also print the version of Python being used.

Updated the library package to 0.3.1.

On branch full-integration
 Changes to be committed:
	modified:   002.library/python/amajusmating/faps_data.py
	modified:   002.library/python/amajusmating/mcmc.py
	modified:   002.library/python/setup.py
	modified:   003.scripts/setup_FAPS_GPS.py
	modified:   005.results/003_mcmc_restrict_kurtosis/check_posterior.R
	modified:   005.results/003_mcmc_restrict_kurtosis/output/chain1.log
	modified:   005.results/003_mcmc_restrict_kurtosis/output/chain1.out
	modified:   005.results/003_mcmc_restrict_kurtosis/output/chain2.log
	modified:   005.results/003_mcmc_restrict_kurtosis/output/chain2.out
	modified:   005.results/003_mcmc_restrict_kurtosis/output/chain3.log
	modified:   005.results/003_mcmc_restrict_kurtosis/output/chain3.out
	modified:   005.results/003_mcmc_restrict_kurtosis/output/chain4.log
	modified:   005.results/003_mcmc_restrict_kurtosis/output/chain4.out
	modified:   005.results/003_mcmc_restrict_kurtosis/run_MCMC.py